toolkit2: removed irrelevant tests+ Test Fixes
git-svn-id: http://svn.terracotta.org/svn/ehcache@5668 b9324663-ca0f-0410-8574-be9b3887307d

+++ branches/ehcache_toolkit2/terracotta/bootstrap/src/main/java/org/terracotta/modules/ehcache/store/CacheConfigChangeBridge.java
@@ -53,12 +53,12 @@ public class CacheConfigChangeBridge implements CacheConfigurationListener, Tool
 
   @Override
   public void timeToIdleChanged(long oldTimeToIdle, long newTimeToIdle) {
-    changeAndNotify(DynamicConfigType.MAX_TTI_SECONDS, newTimeToIdle);
+    changeAndNotify(DynamicConfigType.MAX_TTI_SECONDS, (int) newTimeToIdle);
   }
 
   @Override
   public void timeToLiveChanged(long oldTimeToLive, long newTimeToLive) {
-    changeAndNotify(DynamicConfigType.MAX_TTL_SECONDS, newTimeToLive);
+    changeAndNotify(DynamicConfigType.MAX_TTL_SECONDS, (int) newTimeToLive);
   }
 
   @Override
